requirements.txt
tensorflow>=2.17.0 
gunicorn
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
requests==2.32.4
pandas==2.3.1
numpy==1.24.3
scikit-learn==1.3.0
schedule==1.2.2
python-dateutil==2.9.0.post0
pytz==2025.2
SQLAlchemy==2.0.41
Werkzeug==3.1.3
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0
greenlet==3.2.3
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
joblib==1.5.1
threadpoolctl==3.6.0
scipy==1.11.1
tzdata==2025.2
typing_extensions==4.14.0
six==1.17.0
